refactor: typography.ts를 lib에서 utils로 이동
- lib: 설정/초기화 코드 (axios, storage, responsive)
- utils: 순수 헬퍼 함수 (typography, date)
- 관련 파일 import 경로 수정
