+++ src/features/bus/screens/BusScreen.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/native';
 import { Text, View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
-import { getFontFamily } from '@/shared/lib/typography';
+import { getFontFamily } from '@/shared/utils/typography';
 
 export function BusScreen() {
   return (
