+++ src/features/match/components/MatchingRouteInfo.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { Text, View } from 'react-native';
 import Svg, { Line } from 'react-native-svg';
 
-import { getFontFamily } from '@/shared/lib/typography';
+import { getFontFamily } from '@/shared/utils/typography';
 
 export function MatchingRouteInfo() {
   const theme = useTheme();
